fix: move device capabilities registration during login
